requirements.txt
gspread==6.2.1
pandas>=1.3.0
fpdf>=1.7.2
google-api-python-client>=2.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
docxtpl
pyinstaller
setuptools<81